Move shared thread helper from gui/ to utils/
