+++ src/cli/cli_main.rs
@@ -11,11 +11,11 @@ use mpris_player::PlaybackStatus;
 use crate::core::microphone_thread::microphone_thread;
 use crate::core::processing_thread::processing_thread;
 use crate::core::http_thread::http_thread;
-use crate::utils::mpris_player::{get_player, update_song};
-
 use crate::core::thread_messages::{GUIMessage, MicrophoneMessage};
 
-use crate::gui::main_window::spawn_big_thread;
+use crate::utils::mpris_player::{get_player, update_song};
+use crate::utils::thread::spawn_big_thread;
+
 
 pub fn cli_main(audio_device: Option<&str>) -> Result<(), Box<dyn Error>> {
     glib::MainContext::default().acquire();
